+++ src/main/java/org/ferhat/librarymanagementrestapi/dao/BookBorrowingRepo.java
@@ -3,5 +3,5 @@ package org.ferhat.librarymanagementrestapi.dao;
 import org.ferhat.librarymanagementrestapi.entity.BookBorrowing;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookBorrowingRepo extends JpaRepository<BookBorrowing, Integer> {
+public interface BookBorrowingRepo extends JpaRepository<BookBorrowing, Long> {
 }
