+++ src/main/java/org/ferhat/librarymanagementrestapi/dao/PublisherRepo.java
@@ -3,5 +3,5 @@ package org.ferhat.librarymanagementrestapi.dao;
 import org.ferhat.librarymanagementrestapi.entity.Publisher;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PublisherRepo extends JpaRepository<Publisher, Integer> {
+public interface PublisherRepo extends JpaRepository<Publisher, Long> {
 }
